Update comm.h
ICANON needs to be off to read raw bytes from the serial interface, otherwise certain codes are interpreted as commands 0x04 = EOT for example and interfere with the use of read(). Added a flag to disable canonical mode so that students can do so prior to reading bytes from the fd.